-fix build issue -move RulesPopup to be in component since its shared popup
