+++ src/components/Popup/RulesPopupJsx.tsx
